requirements.txt
grpcio==1.50.0
cffi==1.15.1
Flask==2.2.2
Flask-Security-Too==5.0.2
Flask-Caching==2.0.1
python-dotenv==0.21.0
Flask-SQLAlchemy==3.0.2
Flask-Babel==2.0.0
email-validator==1.3.0
bcrypt==4.0.1
cryptography==38.0.3
PyQRCode==1.2.1
phonenumberslite==8.13.0
zxcvbn==4.4.28
bleach==5.0.1
google-analytics-data==0.14.2
mailjet-rest==1.3.4